+++ projects/ngx-state-service/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - 2024-02-02
+
+### Added
+
+- `select` method was added to `StateService`
+
+### Removed
+
+- `setFnc` and `setDeepFnc` methods were removed. Functional parameters are now
+  supported by `set` and `setDeep` functions
+
 ## [0.1.4] - 2024-01-31
 
 ### Changed